Clean up published documentation
