+++ packages/native-with-source/docs/tests/create-pdfs.js
@@ -4,8 +4,6 @@ var os = require("os");
 var path = require("path");
 var muhammara = require("@muhammara/native-with-source");
 require.cache[require.resolve("@muhammara/native")] = { exports: muhammara };
-var createLowLevelPdf = require("../../../native/docs/examples/create-low-level-pdf");
-var createRecipePdf = require("../../../native/docs/examples/create-recipe-pdf");
 
 describe("Documentation examples", function () {
   var outputDirectory;
@@ -20,22 +18,30 @@ describe("Documentation examples", function () {
 
   it("creates a low-level PDF", function () {
     var outputPath = path.join(outputDirectory, "low-level.pdf");
+    var pdfWriter = muhammara.createWriter(outputPath);
+    var page = pdfWriter.createPage(0, 0, 595, 842);
 
-    createLowLevelPdf(outputPath);
+    pdfWriter.writePage(page);
+    pdfWriter.end();
 
     var reader = muhammara.createReader(outputPath);
     assert.strictEqual(reader.getPagesCount(), 1);
     reader.end();
   });
 
-  it("creates a Recipe PDF", function (done) {
+  it("creates a Recipe PDF", function () {
     var outputPath = path.join(outputDirectory, "recipe.pdf");
-
-    createRecipePdf(outputPath, function () {
-      var reader = muhammara.createReader(outputPath);
-      assert.strictEqual(reader.getPagesCount(), 1);
-      reader.end();
-      done();
+    var Recipe = require("@muhammara/native").Recipe;
+    var pdfDoc = new Recipe("new", outputPath, {
+      version: 1.6,
+      author: "John Doe",
+      title: "A brand new PDF",
     });
+
+    pdfDoc.createPage("letter").endPage().endPDF();
+
+    var reader = muhammara.createReader(outputPath);
+    assert.strictEqual(reader.getPagesCount(), 1);
+    reader.end();
   });
 });
