--- packages/native/docs/examples/create-low-level-pdf.js
@@ -1,11 +1,0 @@
-var muhammara = require("@muhammara/native");
-
-function createLowLevelPdf(outputPath) {
-  var writer = muhammara.createWriter(outputPath);
-  var page = writer.createPage(0, 0, 595, 842);
-
-  writer.writePage(page);
-  writer.end();
-}
-
-module.exports = createLowLevelPdf;
