--- packages/native/docs/examples/replace-recipe-text.js
@@ -1,7 +1,0 @@
-var Recipe = require("@muhammara/native").Recipe;
-
-function replaceRecipeText(inputPath, outputPath) {
-  new Recipe(inputPath, outputPath).replaceText("Before", "After").endPDF();
-}
-
-module.exports = replaceRecipeText;
