+++ packages/native/docs/how-to/index.md
@@ -5,6 +5,3 @@ against the current implementation and identify their verification source.
 
 GitHub issues and discussions are used to identify gaps, not copied as
 documentation.
-
-See [Requested Workflow Backlog](backlog.md) for reviewed requests that need
-additional implementation or tests before publication.
