+++ packages/native/docs/low-level/index.md
@@ -4,4 +4,5 @@ The low-level API exposes the native PDF writer and reader, page content
 contexts, streams, forms, copying contexts, and raw PDF drawing operators.
 
 The complete 15-page HummusJS wiki was reviewed and rewritten into this section.
-Use the pages in this navigation rather than the legacy wiki source material.
+Use the pages in this navigation for supported writer, reader, stream, form,
+copying, and drawing operations rather than the legacy wiki source material.
