+++ packages/native/mkdocs.yml
@@ -99,11 +99,9 @@ nav:
           - Add Content to Rotated Pages: how-to/add-content-to-rotated-pages.md
           - Place and Transform Images: how-to/place-and-transform-images.md
           - Set Page Boxes: how-to/set-page-boxes.md
-          - Requested Workflow Backlog: how-to/backlog.md
       - Project Documentation:
           - Breaking Changes: breaking-changes.md
           - Contributing Documentation: contributing.md
-          - Documentation Sources: provenance.md
   - Project:
       - Development Commands: development.md
       - Security and Vendored Dependencies: security.md
