+++ packages/wasm/CHANGELOG.md
@@ -47,6 +47,8 @@ All notable changes to `@muhammara/wasm` are documented in this file.
 
 - Document the native-only writer events and Wasm-only `dispose()` methods
   [#624](https://github.com/julianhille/MuhammaraJS/issues/624)
+- Document the browser example's grayscale form XObject tab
+  [#630](https://github.com/julianhille/MuhammaraJS/issues/630)
 - Reject invalid page indices and object IDs in the reader's `getPageObjectID()`
   and `getXrefEntry()`, and reject values of 2^32 and above in every reader
   method that takes an index, so both ends fail the same way [#581](https://github.com/julianhille/MuhammaraJS/issues/581)
