+++ packages/wasm/mkdocs.yml
@@ -61,7 +61,6 @@ nav:
       - Add Content to Rotated Pages: how-to/add-content-to-rotated-pages.md
       - Place and Transform Images: how-to/place-and-transform-images.md
       - Set Page Boxes: how-to/set-page-boxes.md
-      - Requested Workflow Backlog: how-to/backlog.md
   - Compatibility: differences.md
   - Project: development.md
   - Breaking Changes: breaking-changes.md
